updates to move tests for race condition issues
