+++ db/revision_cache_test.go
@@ -1666,158 +1666,9 @@ func TestRevCacheOnDemand(t *testing.T) {
 	log.Printf("Calling getRev for %s, %s", docID, revID)
 	rev, err := collection.getRev(ctx, docID, revID, 0, nil)
 	require.Error(t, err)
-	if base.IsEnterpriseEdition() {
-		fmt.Println("here")
-	}
 	require.ErrorContains(t, err, "missing")
 	// returns empty doc rev
 	assert.Equal(t, "", rev.DocID)
-}
-
-func TestLoadActiveDocFromBucketRevCacheChurn(t *testing.T) {
-	base.SkipImportTestsIfNotEnabled(t)
-
-	dbcOptions := DatabaseContextOptions{
-		RevisionCacheOptions: &RevisionCacheOptions{
-			MaxItemCount: 2,
-			ShardCount:   1,
-		},
-	}
-	var wg sync.WaitGroup
-	db, ctx := SetupTestDBWithOptions(t, dbcOptions)
-	defer db.Close(ctx)
-	collection, ctx := GetSingleDatabaseCollectionWithUser(ctx, t, db)
-	docID := "doc1"
-	_, _, err := collection.Put(ctx, docID, Body{"ver": "0"})
-	require.NoError(t, err)
-	wg.Add(1)
-
-	testCtx, testCtxCancel := context.WithCancel(base.TestCtx(t))
-	defer testCtxCancel()
-
-	for i := 0; i < 2; i++ {
-		docID := fmt.Sprintf("extraDoc%d", i)
-		revID, _, err := collection.Put(ctx, docID, Body{"fake": "body"})
-		require.NoError(t, err)
-		go func() {
-			for {
-				select {
-				case <-testCtx.Done():
-					return
-				default:
-					_, err = db.revisionCache.GetWithRev(ctx, docID, revID, collection.GetCollectionID(), RevCacheOmitDelta) //nolint:errcheck
-				}
-			}
-		}()
-	}
-
-	go func() {
-		for i := 0; i < 100; i++ {
-			err = collection.dataStore.Set(docID, 0, nil, []byte(fmt.Sprintf(`{"ver": "%d"}`, i)))
-			require.NoError(t, err)
-			_, err := db.revisionCache.GetActive(ctx, docID, collection.GetCollectionID())
-			if err != nil {
-				break
-			}
-		}
-		wg.Done()
-	}()
-	wg.Wait()
-	require.NoError(t, err)
-}
-
-func TestLoadRequestedRevFromBucketHighChurn(t *testing.T) {
-
-	dbcOptions := DatabaseContextOptions{
-		RevisionCacheOptions: &RevisionCacheOptions{
-			MaxItemCount: 2,
-			ShardCount:   1,
-		},
-	}
-	var wg sync.WaitGroup
-	db, ctx := SetupTestDBWithOptions(t, dbcOptions)
-	defer db.Close(ctx)
-	collection, ctx := GetSingleDatabaseCollectionWithUser(ctx, t, db)
-	docID := "doc1"
-	rev1ID, _, err := collection.Put(ctx, docID, Body{"ver": "0"})
-	require.NoError(t, err)
-	wg.Add(1)
-
-	testCtx, testCtxCancel := context.WithCancel(base.TestCtx(t))
-	defer testCtxCancel()
-
-	for i := 0; i < 2; i++ {
-		docID := fmt.Sprintf("extraDoc%d", i)
-		revID, _, err := collection.Put(ctx, docID, Body{"fake": "body"})
-		require.NoError(t, err)
-		go func() {
-			for {
-				select {
-				case <-testCtx.Done():
-					return
-				default:
-					_, err = db.revisionCache.GetWithRev(ctx, docID, revID, collection.GetCollectionID(), RevCacheOmitDelta) //nolint:errcheck
-				}
-			}
-		}()
-	}
-
-	var getErr error
-	go func() {
-		for i := 0; i < 100; i++ {
-			_, getErr = db.revisionCache.GetWithRev(ctx, docID, rev1ID, collection.GetCollectionID(), true)
-			if getErr != nil {
-				break
-			}
-		}
-		wg.Done()
-	}()
-	wg.Wait()
-	require.NoError(t, getErr)
-}
-
-func TestPutRevHighRevCacheChurn(t *testing.T) {
-
-	dbcOptions := DatabaseContextOptions{
-		RevisionCacheOptions: &RevisionCacheOptions{
-			MaxItemCount: 2,
-			ShardCount:   1,
-		},
-	}
-	var wg sync.WaitGroup
-	db, ctx := SetupTestDBWithOptions(t, dbcOptions)
-	defer db.Close(ctx)
-	collection, ctx := GetSingleDatabaseCollectionWithUser(ctx, t, db)
-	docID := "doc1"
-	wg.Add(1)
-
-	testCtx, testCtxCancel := context.WithCancel(base.TestCtx(t))
-	defer testCtxCancel()
-
-	for i := 0; i < 2; i++ {
-		docID := fmt.Sprintf("extraDoc%d", i)
-		revID, _, err := collection.Put(ctx, docID, Body{"fake": "body"})
-		require.NoError(t, err)
-		go func() {
-			for {
-				select {
-				case <-testCtx.Done():
-					return
-				default:
-					_, err = db.revisionCache.GetWithRev(ctx, docID, revID, collection.GetCollectionID(), RevCacheOmitDelta) //nolint:errcheck
-				}
-			}
-		}()
-	}
-
-	go func() {
-		for i := 0; i < 100; i++ {
-			docRev := DocumentRevision{DocID: docID, RevID: fmt.Sprintf("1-%d", i), CV: &Version{SourceID: "someSrc", Value: uint64(i)}, BodyBytes: []byte(fmt.Sprintf(`{"ver": "%d"}`, i)), History: Revisions{"start": 1}}
-			db.revisionCache.Put(ctx, docRev, collection.GetCollectionID())
-		}
-		wg.Done()
-	}()
-	wg.Wait()
 }
 
 // documentRevisionForCacheTest creates a document revision with the specified body and key, and a hardcoded revID, cv and history:
